More sqllogictest refactoring
